# src/main/resources/v1/data-catalog-openapi.yaml
openapi: 3.0.1
info:
  title: Data Catalog
  description: API for Data Catalog
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: PDLEAMZEUG@pdl.internal.ericsson.com
  license:
    name: COPYRIGHT Ericsson "201"9
    url: https://www.ericsson.com/en/legal
  version: 1.0.0
servers:
- url: //localhost:8080/catalog
...
paths:
...
  /v1/message-schema:
    get:
      tags:
      - /v1/message-schema
      summary: get all messageSchema details or filter by QueryParams
      operationId: queryAllMessageSchemaByParamsUsingGET
      parameters:
      - name: dataCategory
        in: query
        description: DataCategory to which the dataProviderType is associated
        schema:
          type: string
      - name: dataProvider
        in: query
        description: DataProviderType Identifier to which the dataSpace is associated
        schema:
          type: string
      - name: dataSpace
        in: query
        description: Dataspace name to which the dataProviderType is associated
        schema:
          type: string
      responses:
        "200":
          description: get MessageSchema details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageSchema'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: MessageSchema details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
    post:
      tags:
      - /v1/message-schema
      summary: create MessageSchema details
      operationId: registerMessageSchemaUsingPOST
      requestBody:
        description: messageSchemaDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageSchemaDto'
        required: true
      responses:
        "201":
          description: created MessageSchema details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSchemaDto'
        "400":
          description: Bad Request in values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: MessageSchema details exists already
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
      x-codegen-request-body-name: messageSchemaDto
  /v1/message-schema/{id}:
    get:
      tags:
      - /v1/message-schema
      summary: get messageSchema details by id
      operationId: queryMessageSchemaByIdUsingGET
      parameters:
      - name: id
        in: path
        description: id uniquely identifies messageSchema
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get messageSchema details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSchemaDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: messageSchema details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
...
components:
  schemas:
    BulkDataRepository:
      title: BulkDataRepository
      type: object
      properties:
        accessEndpoints:
          type: array
          items:
            type: string
        clusterName:
          type: string
        fileFormat:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
        fileFormatIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        name:
          type: string
        nameSpace:
          type: string
...
    CatalogResponse:
      title: CatalogResponse
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timeStamp:
          type: string
          format: date-time
    DataCollector:
      title: DataCollector
      type: object
      properties:
        collectorId:
          type: string
        controlEndpoint:
          $ref: '#/components/schemas/URL'
        id:
          type: integer
          format: int32
        name:
          type: string
        new:
          type: boolean
...
    DataProviderType:
      title: DataProviderType
      type: object
      properties:
        dataCategory:
          type: string
          enum:
          - CM_NOTIFICATIONS
          - PM_STATS
          - FM_NOTIFICATIONS
          - PM_EVENTS
          - CM_EXPORT
        dataSpace:
          $ref: '#/components/schemas/DataSpace'
        dataSpaceId:
          type: integer
          format: int32
        fileFormatIds:
          type: array
          items:
            type: integer
            format: int32
        fileFormats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
        id:
          type: integer
          format: int32
        messageSchemaIds:
          type: array
          items:
            type: integer
            format: int32
        messageSchemas:
          type: array
          items:
            $ref: '#/components/schemas/MessageSchema'
        new:
          type: boolean
        providerTypeId:
          type: string
        providerVersion:
          type: string
...
    DataSpace:
      title: DataSpace
      type: object
      properties:
        dataProviderTypeIds:
          type: array
          items:
            type: integer
            format: int32
        dataProviderTypes:
          type: array
          items:
            $ref: '#/components/schemas/DataProviderType'
        id:
          type: integer
          format: int32
        name:
          type: string
        new:
          type: boolean
...
    FileFormat:
      title: FileFormat
      type: object
      properties:
        bulkDataRepository:
          $ref: '#/components/schemas/BulkDataRepository'
        bulkDataRepositoryId:
          type: integer
          format: int32
        dataCollector:
          $ref: '#/components/schemas/DataCollector'
        dataEncoding:
          type: string
          enum:
          - JSON
          - CSV
          - XML
          - PARQUET
        dataProviderType:
          $ref: '#/components/schemas/DataProviderType'
        id:
          type: integer
          format: int32
        new:
          type: boolean
        notificationTopic:
          $ref: '#/components/schemas/NotificationTopic'
        reportOutputPeriodList:
          type: array
          items:
            type: integer
            format: int32
        specificationReference:
          type: string
...
    MessageBus:
      title: MessageBus
      type: object
      properties:
        accessEndpoints:
          type: array
          items:
            type: string
        clusterName:
          type: string
        id:
          type: integer
          format: int32
        messageDataTopicIds:
          type: array
          items:
            type: integer
            format: int32
        messageDataTopics:
          type: array
          items:
            $ref: '#/components/schemas/MessageDataTopic'
        messageStatusTopicIds:
          type: array
          items:
            type: integer
            format: int32
        messageStatusTopics:
          type: array
          items:
            $ref: '#/components/schemas/MessageStatusTopic'
        name:
          type: string
        nameSpace:
          type: string
        notificationTopicIds:
          type: array
          items:
            type: integer
            format: int32
        notificationTopics:
          type: array
          items:
            $ref: '#/components/schemas/NotificationTopic'
...
    MessageDataTopic:
      title: MessageDataTopic
      type: object
      properties:
        encoding:
          type: string
          enum:
          - JSON
          - AVRO
          - PROTOBUF
        id:
          type: integer
          format: int32
        messageBus:
          $ref: '#/components/schemas/MessageBus'
        messageBusId:
          type: integer
          format: int32
        messageSchemaIds:
          type: array
          items:
            type: integer
            format: int32
        messageSchemas:
          type: array
          items:
            $ref: '#/components/schemas/MessageSchema'
        messageStatusTopic:
          $ref: '#/components/schemas/MessageStatusTopic'
        messageStatusTopicId:
          type: integer
          format: int32
        name:
          type: string
        new:
          type: boolean
...
    MessageSchema:
      title: MessageSchema
      type: object
      properties:
        dataCollector:
          $ref: '#/components/schemas/DataCollector'
        dataProviderType:
          $ref: '#/components/schemas/DataProviderType'
        dataTopic:
          $ref: '#/components/schemas/MessageDataTopic'
        dataTopicId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        new:
          type: boolean
        specificationReference:
          type: string
    MessageSchemaDto:
      title: MessageSchemaDto
      type: object
      properties:
        dataCollectorId:
          type: string
          format: uuid
        dataProviderTypeId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        messageDataTopicId:
          type: integer
          format: int32
        specificationReference:
          type: string
    MessageStatusTopic:
      title: MessageStatusTopic
      type: object
      properties:
        encoding:
          type: string
          enum:
          - JSON
          - AVRO
          - PROTOBUF
        id:
          type: integer
          format: int32
        messageBus:
          $ref: '#/components/schemas/MessageBus'
        messageBusId:
          type: integer
          format: int32
        messageDataTopicIds:
          type: array
          items:
            type: integer
            format: int32
        messageDataTopics:
          type: array
          items:
            $ref: '#/components/schemas/MessageDataTopic'
        name:
          type: string
        new:
          type: boolean
        specificationReference:
          type: string
...
    NotificationTopic:
      title: NotificationTopic
      type: object
      properties:
        encoding:
          type: string
          enum:
          - JSON
          - AVRO
          - PROTOBUF
        fileFormatIds:
          type: array
          items:
            type: integer
            format: int32
        fileFormats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
        id:
          type: integer
          format: int32
        messageBus:
          $ref: '#/components/schemas/MessageBus'
        messageBusId:
          type: integer
          format: int32
        name:
          type: string
        new:
          type: boolean
        specificationReference:
          type: string
...
    URL:
      title: URL
      type: object
      properties:
        authority:
          type: string
        content:
          type: object
          properties: {}
        defaultPort:
          type: integer
          format: int32
        file:
          type: string
        host:
          type: string
        path:
          type: string
        port:
          type: integer
          format: int32
        protocol:
          type: string
        query:
          type: string
        ref:
          type: string
        userInfo:
          type: string
...